refactor(lib): Refactor library naming
