+++ server/src/express/api/service/logout.ts
@@ -15,7 +15,7 @@ import { LoadType } from '../../../@types/Express.types.js';
 
 import type { Request, Response } from 'express';
 import type { Database } from '../../../lib/database/Maria.js';
-import type { SessionManager } from '../../../lib/SessionManager/SessionManager.js';
+import type { SessionManager } from '../../../lib/session-manager/SessionManager.js';
 import type { ApiConfig } from '../../../@types/Config.types.js';
 import type { ResultData } from "../../../@types/Express.types.js";
