+++ server/src/lib/session-manager/SessionManager.ts
